tvsubs: Make sure we have srt file in zip
When we download a file from tvsubtitles, let's make sure we have a srt
file in the zip file. What we did before was to extract the first file
of the archive. Now, we look for a file that ends with srt or sub. If
none is present, we raise an exception.