tvsubs: Use dice_coef to find best match
TVSubtitles.net might give a lot of answers when looking for
a specific TV Show. We have to filter out and find the correct one
automatically by discarding those which have a different name.
We have a maximum distance here expressed as a coefficient, and
we look for the best match.