misc: Add strings_contained function
This function searches every word from reduced_name in the
complete_name string.
The idea is that, we can make sure that the title we want to validate
is fully contained in the file name.
For example:
If we think the movie is named "The Shawshank Redemption", we will look
for "The", "Shawshank", and "Redemption" in the filename.